refactor(sc): move update smartcontract method
